+++ test/event-01.dart
@@ -2,12 +2,12 @@ import 'package:dquery/dquery.dart';
 
 void main() {
   
-  $document().on('click', (DQueryEvent event) {
+  $document().on('click', (QueryEvent event) {
     print("${event.target}, data:${event.data}");
     
   }, selector: 'div.button');
   
-  $('#trigger').on('click', (DQueryEvent event) {
+  $('#trigger').on('click', (QueryEvent event) {
     print('trigger');
     $('div.button').trigger('click', data: 88);
   });