+++ test/manipulation-02.dart
@@ -6,12 +6,12 @@ void main() {
   final InputElement input = querySelector('#input');
   final ElementQuery $target = $('#target');
   
-  $('#append').on('click', (DQueryEvent event) => $target.append(input.value));
-  $('#prepend').on('click', (DQueryEvent event) => $target.prepend(input.value));
-  $('#after').on('click', (DQueryEvent event) => $target.after(input.value));
-  $('#before').on('click', (DQueryEvent event) => $target.before(input.value));
+  $('#append').on('click', (QueryEvent event) => $target.append(input.value));
+  $('#prepend').on('click', (QueryEvent event) => $target.prepend(input.value));
+  $('#after').on('click', (QueryEvent event) => $target.after(input.value));
+  $('#before').on('click', (QueryEvent event) => $target.before(input.value));
   
-  $('#appendTo').on('click', (DQueryEvent event) => $(input.value).appendTo('#target'));
-  $('#prependTo').on('click', (DQueryEvent event) => $(input.value).prependTo('#target'));
+  $('#appendTo').on('click', (QueryEvent event) => $(input.value).appendTo('#target'));
+  $('#prependTo').on('click', (QueryEvent event) => $(input.value).prependTo('#target'));
   
 }